+++ src/platform_impl/ios/event_loop.rs
@@ -7,6 +7,8 @@ use std::{
     sync::mpsc::{self, Receiver, Sender},
 };
 
+use raw_window_handle::{RawDisplayHandle, UiKitDisplayHandle};
+
 use crate::{
     dpi::LogicalSize,
     event::Event,
@@ -62,6 +64,10 @@ impl<T: 'static> EventLoopWindowTarget<T> {
         let monitor = unsafe { monitor::main_uiscreen() };
 
         Some(RootMonitorHandle { inner: monitor })
+    }
+
+    pub fn raw_display_handle(&self) -> RawDisplayHandle {
+        RawDisplayHandle::UiKit(UiKitDisplayHandle::empty())
     }
 }